Issues with Merge approach and combining arrays in arrays of structs.
The problem is how to differentiate between when to merge and when to
remove array entries.

Also, in some cases the values are not addressable so it's not possible
to actually set the values.

Another limitation is that this could easily break with the injection
of additional containers in StatefulSet/Deployment specs due to 3rd
party integration's such as Istio